[GENERAL] Now the cleaner does not delete the main.pdf files.

+++ Cleaner.py
@@ -14,10 +14,10 @@ import json
 # In[20]:
 
 
-with open("cleaner_settings.json","r") as f:
+with open("cleaner_settings.json", "r") as f:
     settings = json.load(f)
-    
-black_list = settings["black-list"]    
+
+black_list = settings["black-list"]
 white_list = settings["white-list"]
 
 
@@ -28,14 +28,14 @@ def get_file_in_folder(path=None):
     if path == None:
         path = os.getcwd()
     path += "/"
-    file_list   = []
+    file_list = []
     folder_list = []
     for entity in listdir(path):
         if os.path.isdir(path+entity):
             folder_list.append(entity)
         else:
             file_list.append(path+entity)
-            
+
     for folder in folder_list:
         for file in get_file_in_folder(path+folder):
             file_list.append(file)
@@ -52,15 +52,14 @@ for file in get_file_in_folder():
         if white_folder in file:
             skip = True
             continue
-            
+
     if skip == True:
         continue
-    
+
     # get the extension of the file
-    _, extension = os.path.splitext(file)
-    
+    filename, extension = os.path.splitext(file)
+
     # remove blacklisted extension
-    if extension in black_list and extension != "":
+    if extension in black_list and extension != "" and not file.endswith("main.pdf"):
         print(file)
         os.remove(file)
-
